TKSS-963: Move kona header files to include directory
